containers/orchestrator/requirements.txt
Flask==3.0.0
requests==2.31.0
pyyaml==6.0.1
prometheus_client==0.19.0